+++ beakpeek/lib/View/Login/landing_tab_2.dart
@@ -5,8 +5,8 @@ import 'package:beakpeek/LandingText/cust_rich_text.dart';
 import 'package:beakpeek/LandingText/cust_text_span.dart';
 import 'package:flutter/material.dart';
 
-class LandingTab2Light extends StatelessWidget {
-  const LandingTab2Light({super.key});
+class LandingTab2 extends StatelessWidget {
+  const LandingTab2({super.key});
 
   @override
   Widget build(BuildContext context) {
