+++ beakpeek/lib/main.dart
@@ -1,11 +1,10 @@
-import 'package:beakpeek/bird_search.dart';
-import 'package:beakpeek/home.dart';
-import 'package:beakpeek/landing_page.dart';
-import 'package:beakpeek/log_in.dart';
-import 'package:beakpeek/map_info.dart';
-import 'package:beakpeek/sign_up.dart';
+import 'package:beakpeek/View/Home/bird_search.dart';
+import 'package:beakpeek/View/Home/home.dart';
+import 'package:beakpeek/View/Login/landing_page.dart';
+import 'package:beakpeek/View/Login/log_in.dart';
+import 'package:beakpeek/View/Home/map_info.dart';
+import 'package:beakpeek/View/Login/sign_up.dart';
 import 'package:flutter/material.dart';
-
 
 void main() {
   runApp(MaterialApp(
@@ -16,7 +15,7 @@ void main() {
       '/signup': (context) => const SignUp(),
       '/home': (context) => const Home(),
       '/birdSearch': (context) => const BirdSearch(),
-      '/map' : (context) => const MapInfo(),
+      '/map': (context) => const MapInfo(),
     },
   ));
 }
