+++ beakpeek/test/bird_test.dart
@@ -1,8 +1,9 @@
+import 'package:beakpeek/Module/bird.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:http/http.dart' as http;
 import 'package:mockito/mockito.dart';
 import 'package:mockito/annotations.dart';
-import 'package:beakpeek/bird.dart';
+import 'package:beakpeek/View/Home/bird.dart';
 // class MockClient extends Mock implements http.Client {}
 
 import 'bird_test.mocks.dart';
